+++ src/main/kotlin/de/dkfz/odcf/guide/annotation/ExcludeFromComparison.kt
@@ -1,0 +1,5 @@
+package de.dkfz.odcf.guide.annotation
+
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.PROPERTY)
+annotation class ExcludeFromComparison